+++ authority-portal-frontend/src/app/common/layouts/portal-layout/sidebar/sidebar.component.ts
@@ -1,6 +1,7 @@
 import {Component, Inject, OnDestroy, OnInit} from '@angular/core';
 import {Subject} from 'rxjs';
 import {takeUntil} from 'rxjs/operators';
+import {UserAvatarModel} from 'src/app/common/layouts/portal-layout/user-avatar/user-avatar.component';
 import {UserInfo} from '../../../../../../../authority-portal-backend/authority-portal-api-client-ts';
 import {APP_CONFIG, AppConfig} from '../../../../core/config/app-config';
 import {GlobalStateUtils} from '../../../../core/global-state/global-state-utils';
@@ -11,6 +12,7 @@ import {GlobalStateUtils} from '../../../../core/global-state/global-state-utils
 })
 export class SidebarComponent implements OnInit, OnDestroy {
   userInfo: UserInfo | null = null;
+  userAvatarData!: UserAvatarModel;
 
   smallDisplayExpandedMenu: boolean = false;
 
@@ -24,6 +26,11 @@ export class SidebarComponent implements OnInit, OnDestroy {
       .pipe(takeUntil(this.ngOnDestroy$))
       .subscribe((userInfo) => {
         this.userInfo = userInfo;
+        this.userAvatarData = {
+          firstName: userInfo.firstName,
+          lastName: userInfo.lastName,
+          role: userInfo.roles.length ? userInfo.roles[0] : '',
+        };
       });
   }