+++ authority-portal-frontend/src/app/pages/participant-organization-profile-page/state/participant-organization-profile-page-actions.ts
@@ -1,0 +1,10 @@
+const tag = 'ParticipantOrganizationProfilePage';
+
+export class SetOrganizationMdsId {
+  static readonly type = `[${tag}] Set Organization Mds Id`;
+  constructor(public organizationMdsId: string) {}
+}
+
+export class RefreshOrganization {
+  static readonly type = `[${tag}] Refresh Organization`;
+}