+++ authority-portal-frontend/src/app/pages/participant-user-detail-page/participant-user-detail-page.module.ts
@@ -1,27 +1,20 @@
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
-import {RouterModule} from '@angular/router';
 import {NgxsModule} from '@ngxs/store';
 import {ErrorElementModule} from 'src/app/common/components/error-element/error-element.module';
 import {LoadingElementModule} from 'src/app/common/components/loading-element/loading-element.module';
 import {PortalLayoutModule} from 'src/app/common/layouts/portal-layout/portal-layout.module';
-import {MaterialModule} from 'src/app/common/material/material.module';
 import {SharedModule} from 'src/app/shared/shared.module';
-import {AuthorityOrganizationDetailPageComponent} from './authority-organization-detail-page/authority-organization-detail-page.component';
-import {AuthorityOrganizationDetailPageStateImpl} from './state/authority-organization-detail-page-state-impl';
+import {ParticipantUserDetailPageComponent} from './participant-user-detail-page/participant-user-detail-page.component';
+import {ParticipantUserDetailPageStateImpl} from './state/participant-user-detail-page-state-impl';
 
 @NgModule({
-  declarations: [AuthorityOrganizationDetailPageComponent],
+  declarations: [ParticipantUserDetailPageComponent],
   imports: [
-    // Angular
     CommonModule,
-    RouterModule,
-
-    // Angular Material
-    MaterialModule,
 
     // NGXS
-    NgxsModule.forFeature([AuthorityOrganizationDetailPageStateImpl]),
+    NgxsModule.forFeature([ParticipantUserDetailPageStateImpl]),
 
     // Authority Portal
     SharedModule,
@@ -30,4 +23,4 @@ import {AuthorityOrganizationDetailPageStateImpl} from './state/authority-organi
     ErrorElementModule,
   ],
 })
-export class AuthorityOrganizationDetailPageModule {}
+export class ParticipantUserDetailPageModule {}