+++ authority-portal-frontend/src/app/shared/components/user-detail/user-detail.component.ts
@@ -1,0 +1,11 @@
+import {Component, Input} from '@angular/core';
+import {UserDetailDto} from '@sovity.de/authority-portal-client';
+
+@Component({
+  selector: 'app-user-detail',
+  templateUrl: './user-detail.component.html',
+})
+export class UserDetailComponent {
+  @Input() userId!: string;
+  @Input() user!: UserDetailDto;
+}